requirements.txt
selenium~=4.16.0
pytest~=7.4.3
allure-pytest~=2.13.2
